Create CMP and consent strings activity
