+++ hybid.demo/src/main/java/net/pubnative/lite/demo/ui/activities/consent/ConsentStringsActivity.kt
@@ -5,10 +5,10 @@ import androidx.appcompat.app.AppCompatActivity
 import androidx.appcompat.widget.Toolbar
 import net.pubnative.lite.demo.R
 
-class CCPAPrivacyActivity : AppCompatActivity() {
+class ConsentStringsActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_ccpa_privacy)
+        setContentView(R.layout.activity_consent_strings)
         val toolbar = findViewById<Toolbar>(R.id.toolbar)
         setSupportActionBar(toolbar)
         supportActionBar?.setDisplayHomeAsUpEnabled(true)
