--- hybid.demo/src/main/java/net/pubnative/lite/demo/ui/activities/consent/GDPRConsentActivity.kt
@@ -1,22 +1,0 @@
-package net.pubnative.lite.demo.ui.activities.consent
-
-import android.os.Bundle
-import androidx.appcompat.app.AppCompatActivity
-import androidx.appcompat.widget.Toolbar
-import net.pubnative.lite.demo.R
-
-class GDPRConsentActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_gdpr_consent)
-        val toolbar = findViewById<Toolbar>(R.id.toolbar)
-        setSupportActionBar(toolbar)
-        supportActionBar?.setDisplayHomeAsUpEnabled(true)
-        supportActionBar?.setDisplayShowHomeEnabled(true)
-    }
-
-    override fun onSupportNavigateUp(): Boolean {
-        onBackPressed()
-        return true
-    }
-}
