+++ hybid.demo/src/main/res/layout/activity_consent_strings.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="net.pubnative.lite.demo.ui.activities.consent.GDPRConsentActivity">
+    tools:context="net.pubnative.lite.demo.ui.activities.consent.ConsentStringsActivity">
 
     <com.google.android.material.appbar.AppBarLayout
         android:layout_width="match_parent"
@@ -22,10 +22,10 @@
 
     <fragment
         android:id="@+id/fragment"
-        android:name="net.pubnative.lite.demo.ui.fragments.consent.GDPRConsentFragment"
+        android:name="net.pubnative.lite.demo.ui.fragments.consent.ConsentStringsFragment"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         app:layout_behavior="@string/appbar_scrolling_view_behavior"
-        tools:layout="@layout/fragment_gdpr_consent" />
+        tools:layout="@layout/fragment_consent_strings" />
 
 </androidx.coordinatorlayout.widget.CoordinatorLayout>
